+++ yash-syntax/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to `yash-syntax` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.7.0] - Unreleased
+## [0.6.1] - Unreleased
 
 ### Added
 
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `SyntaxError` instead of `InvalidCommandToken` depending on the type of
   erroneous tokens.
 - Dependency versions
-    - Rust 1.58.0 → 1.64.0
+    - Rust 1.58.0 → 1.66.0
 
 ## [0.6.0] - 2022-10-01